+++ src/Bibliophilarr.Api.V1/Books/BookFormatStatusResource.cs
@@ -7,6 +7,7 @@ namespace Bibliophilarr.Api.V1.Books
         public FormatType FormatType { get; set; }
         public bool Monitored { get; set; }
         public bool HasFile { get; set; }
+        public int FileCount { get; set; }
         public int? QualityProfileId { get; set; }
         public string QualityProfileName { get; set; }
     }